Dialogue Object Editing
Adding properties to better define dialogue objects, including specifying the character speaking and expression, alongside having the option to assign an item as a requirement for a response to show. Also, setting up the next dialogue to play once the current dialogue is finished, as well as setting a dialogue object as "special".
Note, these are just added properties and haven't actually been applied to the code yet.